Done: Refactoring UIIconView, UIStickyView
