+++ KifuSF/KFVCollaboratorInfo.swift
@@ -12,7 +12,7 @@ class KFVCollaboratorInfo: UIDescriptorView {
     
     let headlineLabel = UILabel()
     let descriptionLabel = UILabel()
-    let statisticsStickyView = KFVSticky<UIStatisticsView>(stickySide: .top)
+    let statisticsStickyView = UIStickyView<UIStatisticsView>(stickySide: .top)
     
     override func configureLayout() {
         contentsStackView.addArrangedSubview(headlineLabel)
